server/requirements.txt
fastapi
uvicorn[standard]
countrydetails
python-dotenv
google-generativeai
wikipedia
langchain_core
langchain-google-genai
langchain
langchain-community
langchain-openai
googlemaps